fix: proving system id type

+++ aggregation_mode/batcher/src/server/http.rs
@@ -177,7 +177,7 @@ impl BatcherServer {
             .db
             .insert_task(
                 &recovered_address,
-                AggregationModeProvingSystem::SP1.as_u16() as i64,
+                AggregationModeProvingSystem::SP1.as_u16() as i32,
                 &proof_content,
                 &vk_content,
                 None,
